+++ lib/Parser/Alloc.h
@@ -10,10 +10,10 @@ NoReleaseAllocator - allocator that never releases until it is destroyed
 class NoReleaseAllocator
 {
 public:
-    NoReleaseAllocator(long cbFirst = 256, long cbMax = 0x4000 /*16K*/);
+    NoReleaseAllocator(int32 cbFirst = 256, int32 cbMax = 0x4000 /*16K*/);
     ~NoReleaseAllocator(void) { FreeAll(); }
 
-    void *Alloc(long cb);
+    void *Alloc(int32 cb);
     void FreeAll();
     void Clear() { FreeAll(); }
 
@@ -24,16 +24,16 @@ private:
         // ... DATA ...
     };
     NraBlock * m_pblkList;
-    long m_ibCur;
-    long m_ibMax;
-    long m_cbMinBlock;
-    long m_cbMaxBlock;
+    int32 m_ibCur;
+    int32 m_ibMax;
+    int32 m_cbMinBlock;
+    int32 m_cbMaxBlock;
 
 #if DEBUG
-    long m_cbTotRequested;    // total bytes requested
-    long m_cbTotAlloced;    // total bytes allocated including headers
-    long m_cblk;            // number of blocks including big blocks
-    long m_cpvBig;            // each generates its own big block
-    long m_cpvSmall;        // put in a common block
+    int32 m_cbTotRequested;    // total bytes requested
+    int32 m_cbTotAlloced;    // total bytes allocated including headers
+    int32 m_cblk;            // number of blocks including big blocks
+    int32 m_cpvBig;            // each generates its own big block
+    int32 m_cpvSmall;        // put in a common block
 #endif //DEBUG
 };
