+++ lib/Runtime/Debug/DebugManager.cpp
@@ -155,8 +155,8 @@ namespace Js
         DynamicObject* consoleScope = this->GetConsoleScope(scriptContext);
         Js::RecyclableObject* recyclableObject = Js::RecyclableObject::FromVar(copyFromScope);
 
-        ulong newPropCount = recyclableObject->GetPropertyCount();
-        for (ulong i = 0; i < newPropCount; i++)
+        uint32 newPropCount = recyclableObject->GetPropertyCount();
+        for (uint32 i = 0; i < newPropCount; i++)
         {
             Js::PropertyId propertyId = recyclableObject->GetPropertyId((Js::PropertyIndex)i);
             // For deleted properties we won't have a property id
