+++ lib/Runtime/Library/UriHelper.cpp
@@ -54,9 +54,9 @@ namespace Js
     // the number of characters in the output array.
     // This routine assumes that it's input 'uVal' is a valid Unicode code-point value
     // and does no error checking.
-    ulong UriHelper::ToUTF8( ulong uVal, BYTE bUTF8[MaxUTF8Len])
+    uint32 UriHelper::ToUTF8( uint32 uVal, BYTE bUTF8[MaxUTF8Len])
     {
-        ulong uRet;
+        uint32 uRet;
         if( uVal <= 0x007F )
         {
             bUTF8[0] = (BYTE)uVal;
@@ -64,8 +64,8 @@ namespace Js
         }
         else if( uVal <= 0x07FF )
         {
-            ulong z = uVal & 0x3F;
-            ulong y = uVal >> 6;
+            uint32 z = uVal & 0x3F;
+            uint32 y = uVal >> 6;
             bUTF8[0] = (BYTE) (0xC0 | y);
             bUTF8[1] = (BYTE) (0x80 | z);
             uRet = 2;
@@ -73,9 +73,9 @@ namespace Js
         else if( uVal <= 0xFFFF )
         {
             Assert( uVal <= 0xD7FF || uVal >= 0xE000 );
-            ulong z = uVal & 0x3F;
-            ulong y = (uVal >> 6) & 0x3F;
-            ulong x = (uVal >> 12);
+            uint32 z = uVal & 0x3F;
+            uint32 y = (uVal >> 6) & 0x3F;
+            uint32 x = (uVal >> 12);
             bUTF8[0] = (BYTE) (0xE0 | x);
             bUTF8[1] = (BYTE) (0x80 | y);
             bUTF8[2] = (BYTE) (0x80 | z);
@@ -83,10 +83,10 @@ namespace Js
         }
         else
         {
-            ulong z = uVal & 0x3F;
-            ulong y = (uVal >> 6) &0x3F;
-            ulong x = (uVal >> 12) &0x3F;
-            ulong w = (uVal >> 18);
+            uint32 z = uVal & 0x3F;
+            uint32 y = (uVal >> 6) &0x3F;
+            uint32 x = (uVal >> 12) &0x3F;
+            uint32 w = (uVal >> 18);
             bUTF8[0] = (BYTE) (0xF0 | w);
             bUTF8[1] = (BYTE) (0x80 | x);
             bUTF8[2] = (BYTE) (0x80 | y);
@@ -101,7 +101,7 @@ namespace Js
     // array 'bUTF8'. uLen is the number of characters in the UTF-8 encoding.
     // This routine assumes that a valid UTF-8 encoding of a character is passed in
     // and does no error checking.
-    unsigned long UriHelper::FromUTF8( BYTE bUTF8[MaxUTF8Len], ulong uLen )
+    uint32 UriHelper::FromUTF8( BYTE bUTF8[MaxUTF8Len], uint32 uLen )
     {
         Assert( 1 <= uLen && uLen <= MaxUTF8Len );
         if( uLen == 1 )
@@ -126,16 +126,16 @@ namespace Js
     // The Encode algorithm described in sec. 15.1.3 of the spec. The input string is
     // 'pSz' and the Unescaped set is described by the flags 'unescapedFlags'. The
     // output is a string var.
-    Var UriHelper::Encode(__in_ecount(len) const  char16* pSz, ulong len, unsigned char unescapedFlags, ScriptContext* scriptContext )
+    Var UriHelper::Encode(__in_ecount(len) const  char16* pSz, uint32 len, unsigned char unescapedFlags, ScriptContext* scriptContext )
     {
         BYTE bUTF8[MaxUTF8Len];
 
         // pass 1 calculate output length and error check
-        ulong outputLen = 0;
-        for( ulong k = 0; k < len; k++ )
+        uint32 outputLen = 0;
+        for( uint32 k = 0; k < len; k++ )
         {
             char16 c = pSz[k];
-            ulong uVal;
+            uint32 uVal;
             if( InURISet(c, unescapedFlags) )
             {
                 outputLen = UInt32Math::Add(outputLen, 1);
@@ -148,7 +148,7 @@ namespace Js
                 }
                 else if( c < 0xD800 || c > 0xDBFF )
                 {
-                    uVal = (ulong)c;
+                    uVal = (uint32)c;
                 }
                 else
                 {
@@ -165,7 +165,7 @@ namespace Js
                     }
                     uVal = (c - 0xD800) * 0x400 + (c1 - 0xDC00) + 0x10000;
                 }
-                ulong utfLen = ToUTF8(uVal, bUTF8);
+                uint32 utfLen = ToUTF8(uVal, bUTF8);
                 utfLen = UInt32Math::Mul(utfLen, 3);
                 outputLen = UInt32Math::Add(outputLen, utfLen);
             }
@@ -177,10 +177,10 @@ namespace Js
         char16* outURI = RecyclerNewArrayLeaf(scriptContext->GetRecycler(), char16, allocSize);
         char16* outCurrent = outURI;
 
-        for( ulong k = 0; k < len; k++ )
+        for( uint32 k = 0; k < len; k++ )
         {
             char16 c = pSz[k];
-            ulong uVal;
+            uint32 uVal;
             if( InURISet(c, unescapedFlags) )
             {
                 __analysis_assume(outCurrent < outURI + allocSize);
@@ -196,7 +196,7 @@ namespace Js
 #endif
                 if( c < 0xD800 || c > 0xDBFF )
                 {
-                    uVal = (ulong)c;
+                    uVal = (uint32)c;
                 }
                 else
                 {
@@ -219,8 +219,8 @@ namespace Js
                     uVal = (c - 0xD800) * 0x400 + (c1 - 0xDC00) + 0x10000;
                 }
 
-                ulong utfLen = ToUTF8(uVal, bUTF8);
-                for( ulong j = 0; j < utfLen; j++ )
+                uint32 utfLen = ToUTF8(uVal, bUTF8);
+                for( uint32 j = 0; j < utfLen; j++ )
                 {
 #pragma prefast(suppress: 26014, "buffer length was calculated earlier");
                     swprintf_s(outCurrent, 4, _u("%%%02X"), (int)bUTF8[j] );
@@ -263,19 +263,19 @@ namespace Js
     // The Decode algorithm described in sec. 15.1.3 of the spec. The input string is
     // 'pSZ' and the Reserved set is described by the flags 'reservedFlags'. The
     // output is a string var.
-    Var UriHelper::Decode(__in_ecount(len) const char16* pSz, ulong len, unsigned char reservedFlags, ScriptContext* scriptContext)
+    Var UriHelper::Decode(__in_ecount(len) const char16* pSz, uint32 len, unsigned char reservedFlags, ScriptContext* scriptContext)
     {
         char16 c1;
         char16 c;
         // pass 1 calculate output length and error check
-        ulong outputLen = 0;
-        for( ulong k = 0; k < len; k++ )
+        uint32 outputLen = 0;
+        for( uint32 k = 0; k < len; k++ )
         {
             c = pSz[k];
 
             if( c == '%')
             {
-                ulong start = k;
+                uint32 start = k;
                 if( k + 2 >= len )
                 {
                     JavascriptError::ThrowURIError(scriptContext, JSERR_URIDecodeError /* TODO-ERROR: _u("NEED MESSAGE") */);
@@ -339,7 +339,7 @@ namespace Js
                         bOctets[j] = b;
                     }
 
-                    ulong uVal = UriHelper::FromUTF8( bOctets, n );
+                    uint32 uVal = UriHelper::FromUTF8( bOctets, n );
 
                     if( uVal >= 0xD800 && uVal <= 0xDFFF)
                     {
@@ -381,12 +381,12 @@ namespace Js
         char16* outCurrent = outURI;
 
 
-        for( ulong k = 0; k < len; k++ )
+        for( uint32 k = 0; k < len; k++ )
         {
             c = pSz[k];
             if( c == '%')
             {
-                ulong start = k;
+                uint32 start = k;
 #if DBG
                 Assert(!(k + 2 >= len));
                 if( k + 2 >= len )
@@ -474,7 +474,7 @@ namespace Js
                         bOctets[j] = b;
                     }
 
-                    ulong uVal = UriHelper::FromUTF8( bOctets, n );
+                    uint32 uVal = UriHelper::FromUTF8( bOctets, n );
 
 #if DBG
                     Assert(!(uVal >= 0xD800 && uVal <= 0xDFFF));
@@ -498,8 +498,8 @@ namespace Js
 #endif
                     else
                     {
-                        ulong l = (( uVal - 0x10000) & 0x3ff) + 0xdc00;
-                        ulong h = ((( uVal - 0x10000) >> 10) & 0x3ff) + 0xd800;
+                        uint32 l = (( uVal - 0x10000) & 0x3ff) + 0xdc00;
+                        uint32 h = ((( uVal - 0x10000) >> 10) & 0x3ff) + 0xd800;
 
                         __analysis_assume(outCurrent + 2 <= outURI + allocSize);
                         *outCurrent++ = (char16)h;
